Style/SafeNavigation: disable
Standard enables this rule but i'm recommending we don't.

When a chain of calls needs to guarded against a `nil`, it can often be
meaningful to explicitly highlight which part of that might nil. While
the first `&.` makes this evident, i find the explicit form to be easier
to spot. It is also a bit more cumbersome, which hopefully creates some
motivation to consider refactoring away from the nil-guarded call chain.

In particular, code such as the following would lose significant meaning
if corrected according to this rule:
```
context.user.destroy! if context.user
context.user&.destroy!
``
Note that disabling this cop doesn't disallow the use of `&.`; authors
are still free to use it as they see fit.

https://docs.rubocop.org/rubocop/cops_style.html#stylesafenavigation