Cache webpacker:compile artifacts in CI build
Uses a cache key that includes a custom checksum of our js source.
We first generate md5 hashes of all files under app/javascript, storing
them in a temporary file. The checksum of this file is then appended to
the cache key. This should ensure that webpacker:compile quickly exits
if nothing has changed.